Add 5 second timeout for API requests
Motivation
----------

There was fairly bad API performance regression as we paged through the
resultset. Somewhere around 100k in requests started taking much longer.
Around 2 million or so API requests would regularly begin to timeout on
the server side. Given that we only request a 100 records at a time taking
60 seconds per 100 was unnacceptable. This takes a hybrid approach where
as long as requests are coming in within 5 seconds we're happy but after
that we restart pagination.